requirements.txt
# Add your project requirements here 
# For example: 
fastapi 
uvicorn[standard] 
streamlit 
python-dotenv  
pypdf 
pytesseract 
Pillow 
unstructured 
tiktoken 
pydantic_settings
langchain-huggingface
langchain-chroma
chromadb
sentence-transformers
fitz
tools
python-multipart
requests
langchain-text-splitters
langchain-community